backend/requirements.txt
aiofiles==24.1.0
anyio==4.10.0
bcrypt==3.2.2
certifi==2025.8.3
cffi==2.0.0
click==8.2.1
cloudinary==1.44.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.95.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.3.1
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.7
pymongo==4.6.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.14.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.22.0
google-auth>=2.20.0
google-api-python-client>=2.70.0
aiosmtplib==5.0.0
gunicorn==21.2.0